Moved Pointer to glue.utils
